requirements.txt
streamlit
akshare
pandas
tavily-python
openai
plotly